Change the default font to Menlo and use a system font as fallback
Courier fails to render "mc" properly, even with the height rounding applied in cellDimension.
Not sure if the height rounding is completely the right way to fix the pixel gap that we see without the rounding, but this at least masks the issue a little :)

I think the ultimate fix is to use a background color for each cell based on the attributes of the char in that cell and fill the cell completely.